some old changes i left behind

+++ app/Http/Controllers/MovieController.php
@@ -13,7 +13,7 @@ class MovieController extends Controller
     public function random()
     {
         return view('randommovie', [
-            'movie' => Movie::inRandomOrder()->where('watched','=',0)->first()
+            'movie' => Movie::inRandomOrder()->where('watched', '=', 0)->first()
         ]);
     }
 
@@ -62,7 +62,7 @@ class MovieController extends Controller
             'effort' => request('effort')
         ]);
 
-        return redirect('/addmovie')->with('status', 'Movie added!');
+        return redirect('/addmovie')->with('success', 'Movie added!');
     }
 
     public function edit(Movie $movie)
@@ -75,7 +75,7 @@ class MovieController extends Controller
 
         $attributes = request()->validate([
             'name' => 'required|string|max:255',
-            'image' => 'image|mimes:jpeg,gif,svg,png,jpg|max:2048',
+            'image' => 'image|mimes:jpeg,gif,svg,png,jpg|max:2048|nullable',
             'genre_id' => ['required', Rule::exists('genres', 'id')],
             'releaseyear' => 'digits:4|integer|min:1900|nullable|max:' . (date('Y') + 1),
             'runtime' => 'numeric|nullable',
@@ -91,8 +91,15 @@ class MovieController extends Controller
         }
 
         if (isset($attributes['image'])) {
+            //delete old image path
             $old_image_path = public_path('images/movies') . '/' . $movie->image;
-            unlink($old_image_path);
+            $default = 'no-photo-available.png';
+            //dont delete if default image
+            if (strpos($old_image_path, $default) == true) {
+                unlink($old_image_path);
+            }
+
+            //update new image
             $imageName = time() . '.' . $request->image->extension();
             $attributes['image'] = request()->file('image')->move(public_path('images/movies'), $imageName);
         } else {
@@ -117,13 +124,13 @@ class MovieController extends Controller
     {
         $movie = Movie::findOrFail($id);
 
-         //check image path, dont delete if default
-         $old_image_path = public_path('images/movies') . '/' . $movie->image;
-         if ($movie->image == 'no-photo-available.png') {
-             //do nothing
-         } else {
-             unlink($old_image_path);
-         }
+        //check image path, dont delete if default
+        $old_image_path = public_path('images/movies') . '/' . $movie->image;
+        if ($movie->image == 'no-photo-available.png') {
+            //do nothing
+        } else {
+            unlink($old_image_path);
+        }
         $movie->delete();
 
         return back()->with('danger', 'Movie Deleted!');
